Renamed logging.py to api_logging.py
